Project5/pyproject.toml
[tool.poetry]
name = "project5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.3"
gspread = "^5.12.0"
python-dotenv = "^1.0.0"
logging = "^0.4.9.6"
config = "^0.5.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
